Handle child processes with SA_NOCLDWAIT
